Move validate_statistics WS API to recorder
