+++ homeassistant/components/history/__init__.py
@@ -17,7 +17,6 @@ from homeassistant.components.recorder import history, models as history_models
 from homeassistant.components.recorder.statistics import (
     list_statistic_ids,
     statistics_during_period,
-    validate_statistics,
 )
 from homeassistant.components.recorder.util import session_scope
 from homeassistant.const import (
@@ -110,7 +109,6 @@ async def async_setup(hass, config):
         ws_get_statistics_during_period
     )
     hass.components.websocket_api.async_register_command(ws_get_list_statistic_ids)
-    hass.components.websocket_api.async_register_command(ws_validate_statistics)
 
     return True
 
@@ -178,23 +176,6 @@ async def ws_get_list_statistic_ids(
         list_statistic_ids,
         hass,
         msg.get("statistic_type"),
-    )
-    connection.send_result(msg["id"], statistic_ids)
-
-
-@websocket_api.websocket_command(
-    {
-        vol.Required("type"): "history/validate_statistics",
-    }
-)
-@websocket_api.async_response
-async def ws_validate_statistics(
-    hass: HomeAssistant, connection: websocket_api.ActiveConnection, msg: dict
-) -> None:
-    """Fetch a list of available statistic_id."""
-    statistic_ids = await hass.async_add_executor_job(
-        validate_statistics,
-        hass,
     )
     connection.send_result(msg["id"], statistic_ids)
